requirements.txt
streamlit
streamlit-webrtc
opencv-python
tensorflow
keras
pillow
numpy
matplotlib
pandas
seaborn
scikit-learn
scipy
facenet-pytorch
ultralytics
tqdm
rich
markdown-it-py
mdurl
pygments
protobuf

